Add files via upload
I am not sure how many heard Professor F. William Lawvere calling for a serious study of constitutive relations, beginning with this 1978 notes on thermostatics and going all the way to his relatively recent 2007 axiomatization of cohesion.  I was fortunate to learn of Professor F. William Lawvere's passion, reason, and conviction of the value of category theory/functorial semantics in the advancement of material sciences.  Unfortunately, for whatever reason, I didn't get to work on it, which makes me all the more jealous ;) seeing my good friend: The Prophet of Köningsberg abstract away cardinality to study constituent relations all on his own, totally unaware of Professor F. William Lawvere's work in this regard (https://zenodo.org/records/4039859, https://zenodo.org/records/4039873); speak of great minds think alike :)  Admittedly, I am not a member of the society of great minds, but I have two good friends with great minds, which is a source of great joy and pride for me :)